add parse_test
