+++ bazel/execution/client_test.go
@@ -4,15 +4,12 @@ package execution
 //go:generate mockgen -destination=mock_longrunning/opclient_mock.go google.golang.org/genproto/googleapis/longrunning OperationsClient
 
 import (
-	"reflect"
 	"testing"
 
 	"github.com/golang/mock/gomock"
-	"github.com/golang/protobuf/ptypes/any"
 	"github.com/golang/protobuf/ptypes/empty"
 	"golang.org/x/net/context"
 	"google.golang.org/genproto/googleapis/longrunning"
-	"google.golang.org/genproto/googleapis/rpc/status"
 	"google.golang.org/grpc"
 
 	"github.com/twitter/scoot/bazel/execution/mock_longrunning"
@@ -106,64 +103,6 @@ func TestClientExecute(t *testing.T) {
 	_, _, _, err = ParseExecuteOperation(op)
 	if err != nil {
 		t.Fatalf("Error parsing resulting Operation: %s", err)
-	}
-}
-
-func TestExtractOpFromJsonError(t *testing.T) {
-	opBytes := []byte(`{"name":"testName","metadata":{"type_url":"type.googleapis.com/build.bazel.remote.execution.v2.ExecuteOperationMetadata","value":"testVal"},"done":true,"Result":{"Error":{"code":1,"message":"CANCELLED"}}}`)
-
-	metadata := &any.Any{
-		TypeUrl: "type.googleapis.com/build.bazel.remote.execution.v2.ExecuteOperationMetadata",
-		Value:   []byte("testVal"),
-	}
-	opErr := &longrunning.Operation_Error{
-		Error: &status.Status{
-			Code:    int32(1),
-			Message: "CANCELLED",
-		},
-	}
-	expOp := &longrunning.Operation{
-		Name:     "testName",
-		Metadata: metadata,
-		Done:     true,
-		Result:   opErr,
-	}
-
-	gotOp, err := ExtractOpFromJson(opBytes)
-	if err != nil {
-		t.Fatalf("Received error extracting operation from json: %s", err)
-	}
-	if !reflect.DeepEqual(gotOp, expOp) {
-		t.Fatalf("Expected gotOp to equal expOp.\ngotOp: %+v\nexpOp: %+v", gotOp, expOp)
-	}
-}
-
-func TestExtractOpFromJsonResponse(t *testing.T) {
-	opBytes := []byte(`{"name":"testName","metadata":{"type_url":"type.googleapis.com/build.bazel.remote.execution.v2.ExecuteOperationMetadata","value":"testVal"},"done":true,"Result":{"Response":{"type_url":"type.googleapis.com/build.bazel.remote.execution.v2.ExecuteOperationMetadata","value":"testVal"}}}`)
-
-	metadata := &any.Any{
-		TypeUrl: "type.googleapis.com/build.bazel.remote.execution.v2.ExecuteOperationMetadata",
-		Value:   []byte("testVal"),
-	}
-	opResp := &longrunning.Operation_Response{
-		Response: &any.Any{
-			TypeUrl: "type.googleapis.com/build.bazel.remote.execution.v2.ExecuteOperationMetadata",
-			Value:   []byte("testVal"),
-		},
-	}
-	expOp := &longrunning.Operation{
-		Name:     "testName",
-		Metadata: metadata,
-		Done:     true,
-		Result:   opResp,
-	}
-
-	gotOp, err := ExtractOpFromJson(opBytes)
-	if err != nil {
-		t.Fatalf("Received error extracting operation from json: %s", err)
-	}
-	if !reflect.DeepEqual(gotOp, expOp) {
-		t.Fatalf("Expected gotOp to equal expOp.\ngotOp: %+v\nexpOp: %+v", gotOp, expOp)
 	}
 }
